Extract the ANSI state machine into a separate file.
